Change float time tick event

+++ Assets/Scripts/Checkers/UI/Views/Base/IPauseWindow.cs
@@ -2,6 +2,6 @@
 
 namespace Checkers.UI.Views.Base {
     public interface IPauseWindow : IWindowView {
-        void SetPauseTime(int time);
+        void SetPauseTime(float time);
     }
 }
